+++ R/lazyframe__lazy.R
@@ -242,6 +242,17 @@ LazyFrame_with_columns = function(...) {
 #' @docType NULL
 LazyFrame_with_column = "use_extendr_wrapper"
 
+#' @title Lazy with_row_count
+#' @description Add a new column at index 0 that counts the rows
+#' @keywords LazyFrame
+#' @param name string name of the created column
+#' @param offset positive integer offset for the start of the counter
+#' @return A new `LazyFrame` object with a counter column in front
+#' @docType NULL
+LazyFrame_with_row_count = function(name, offset = NULL) {
+  .pr$LazyFrame$with_row_count(self, name, offset) |> unwrap()
+}
+
 #' @title Apply filter to LazyFrame
 #' @description Filter rows with an Expression defining a boolean column
 #' @keywords LazyFrame
@@ -924,3 +935,25 @@ LazyFrame_dtypes = method_as_property(function() {
     result() |>
     unwrap("in $dtypes()")
 })
+
+#' @title Explode the DataFrame to long format by exploding the given columns
+#' @keywords LazyFrame
+#'
+#' @param columns Column(s) to be exploded. `Into<Expr>`, list of `Into<Expr>` or a char vec.
+#' Only columns of DataType `List` or `Utf8` can be exploded.
+#' @param ... More columns to explode as above but provided as separate arguments
+#'
+#' @return LazyFrame
+#' @examples
+#' df = pl$LazyFrame(
+#'   letters = c("a", "a", "b", "c"),
+#'   numbers = list(1, c(2, 3), c(4, 5), c(6, 7, 8))
+#' )
+#' df
+#'
+#' df$explode("numbers")$collect()
+LazyFrame_explode = function(columns = list(), ...) {
+  dotdotdot_args = list2(...)
+  .pr$LazyFrame$explode(self, columns, dotdotdot_args) |>
+    unwrap("in explode():")
+}